finalize bader_caller cleanup
Minor code clarifications for bader_caller. Also adds new get_partial_chage() convenience function
