+++ pymatgen/command_line/bader_caller.py
@@ -105,8 +105,6 @@ class BaderAnalysis:
         chgref_filename=None,
         parse_atomic_densities=False,
         cube_filename=None,
-        run_bader=True,
-        bader_output_path=None,
     ):
         """
         Initializes the Bader caller.
@@ -118,10 +116,6 @@ class BaderAnalysis:
             parse_atomic_densities (bool): Optional. turns on atomic partition of the charge density
                 charge densities are atom centered
             cube_filename (str): Optional. The filename of the cube file.
-            run_bader (bool): Optional. Whether to run bader. Defaults to True.
-                If False, BaderAnalysis will look for an ACF.dat file.
-            bader_output_path (str): Path to the bader output files if
-                run_bader is False. Default is the current working directory.
         """
         if not BADEREXE:
             raise RuntimeError(
@@ -134,14 +128,12 @@ class BaderAnalysis:
             raise ValueError("You must provide either a cube file or a CHGCAR")
         if cube_filename and chgcar_filename:
             raise ValueError("You cannot parse a cube and a CHGCAR at the same time.")
-        if bader_output_path is None:
-            bader_output_path = ""
         self.parse_atomic_densities = parse_atomic_densities
 
         if chgcar_filename:
             fpath = os.path.abspath(chgcar_filename)
             self.is_vasp = True
-            self.chgcar = Chgcar.from_file(chgcar_filename)
+            self.chgcar = Chgcar.from_file(fpath)
             self.structure = self.chgcar.structure
             self.potcar = Potcar.from_file(potcar_filename) if potcar_filename is not None else None
             self.natoms = self.chgcar.poscar.natoms
@@ -166,38 +158,37 @@ class BaderAnalysis:
             self.reference_used = bool(chgref_filename)
 
         with ScratchDir("."):
-            if run_bader:
-                tmpfile = "CHGCAR" if chgcar_filename else "CUBE"
-                with zopen(fpath, "rt") as f_in:
-                    with open(tmpfile, "wt") as f_out:
+            tmpfile = "CHGCAR" if chgcar_filename else "CUBE"
+            with zopen(fpath, "rt") as f_in:
+                with open(tmpfile, "wt") as f_out:
+                    shutil.copyfileobj(f_in, f_out)
+            args = [BADEREXE, tmpfile]
+            if chgref_filename:
+                with zopen(chgrefpath, "rt") as f_in:
+                    with open("CHGCAR_ref", "wt") as f_out:
                         shutil.copyfileobj(f_in, f_out)
-                args = [BADEREXE, tmpfile]
-                if chgref_filename:
-                    with zopen(chgrefpath, "rt") as f_in:
-                        with open("CHGCAR_ref", "wt") as f_out:
-                            shutil.copyfileobj(f_in, f_out)
-                    args += ["-ref", "CHGCAR_ref"]
-                if parse_atomic_densities:
-                    args += ["-p", "all_atom"]
-                with subprocess.Popen(args, stdout=subprocess.PIPE, stdin=subprocess.PIPE, close_fds=True) as rs:
-                    stdout, _ = rs.communicate()
-                if rs.returncode != 0:
-                    raise RuntimeError(
-                        "bader exited with return code %d. Please check your bader installation." % rs.returncode
-                    )
+                args += ["-ref", "CHGCAR_ref"]
+            if parse_atomic_densities:
+                args += ["-p", "all_atom"]
+            with subprocess.Popen(args, stdout=subprocess.PIPE, stdin=subprocess.PIPE, close_fds=True) as rs:
+                stdout, _ = rs.communicate()
+            if rs.returncode != 0:
+                raise RuntimeError(
+                    "bader exited with return code %d. Please check your bader installation." % rs.returncode
+                )
 
-                try:
-                    self.version = float(stdout.split()[5])
-                except ValueError:
-                    self.version = -1  # Unknown
-                if self.version < 1.0:
-                    warnings.warn(
-                        "Your installed version of Bader is outdated, calculation of vacuum charge may be incorrect.",
-                        UserWarning,
-                    )
+            try:
+                self.version = float(stdout.split()[5])
+            except ValueError:
+                self.version = -1  # Unknown
+            if self.version < 1.0:
+                warnings.warn(
+                    "Your installed version of Bader is outdated, calculation of vacuum charge may be incorrect.",
+                    UserWarning,
+                )
 
             data = []
-            with open(os.path.join(bader_output_path, "ACF.dat")) as f:
+            with open("ACF.dat") as f:
                 raw = f.readlines()
                 headers = ("x", "y", "z", "charge", "min_dist", "atomic_vol")
                 raw.pop(0)
@@ -346,13 +337,12 @@ class BaderAnalysis:
     def get_oxidation_state_decorated_structure(self, nelects=None):
         """
         Returns an oxidation state decorated structure based on bader analysis results.
+        Each site is assigned a charge based on the computed partial atomic charge from bader.
 
         Note, this assumes that the Bader analysis was correctly performed on a file
         with electron densities
         """
-        charges = [
-            -self.get_charge_transfer(i, None if not nelects else nelects[i]) for i in range(len(self.structure))
-        ]
+        charges = [self.get_partial_charge(i, None if not nelects else nelects[i]) for i in range(len(self.structure))]
         struc = self.structure.copy()
         struc.add_oxidation_state_by_site(charges)
         return struc
@@ -458,7 +448,7 @@ class BaderAnalysis:
 
         chgcar_filename = _get_filepath("CHGCAR")
         if chgcar_filename is None:
-            raise OSError("Could not find CHGCAR!")
+            raise FileNotFoundError("Could not find CHGCAR!")
         potcar_filename = _get_filepath("POTCAR")
         aeccar0 = _get_filepath("AECCAR0")
         aeccar2 = _get_filepath("AECCAR2")
@@ -474,29 +464,6 @@ class BaderAnalysis:
             potcar_filename=potcar_filename,
             chgref_filename=chgref_filename,
         )
-
-
-def get_filepath(filename, warning, path, suffix):
-    """
-    Args:
-        filename: Filename
-        warning: Warning message
-        path: Path to search
-        suffix: Suffixes to search.
-    """
-    paths = glob.glob(os.path.join(path, filename + suffix + "*"))
-    if not paths:
-        warnings.warn(warning)
-        return None
-    if len(paths) > 1:
-        # using reverse=True because, if multiple files are present,
-        # they likely have suffixes 'static', 'relax', 'relax2', etc.
-        # and this would give 'static' over 'relax2' over 'relax'
-        # however, better to use 'suffix' kwarg to avoid this!
-        paths.sort(reverse=True)
-        warnings.warn(f"Multiple files detected, using {os.path.basename(path)}")
-    path = paths[0]
-    return path
 
 
 def bader_analysis_from_path(path, suffix=""):
